+++ R/peaks-functions.R
@@ -224,6 +224,20 @@ NULL
 #'   between `x` and `y` are returned by `type = "inner"`, i.e. only
 #'   peaks present in both spectra are reported.
 #'
+#' - `joinPeaksGnps` matches/maps peaks between spectra with the same approach
+#'   used in GNPS: peaks are considered matching if a) the
+#'   difference in their m/z values is smaller than defined by `tolerance`
+#'   and `ppm` (this is the same as `joinPeaks`) **and** b) the difference of
+#'   their m/z *adjusted* for the difference of the spectras' precursor is
+#'   smaller than defined by `tolerance` and `ppm`. Based on this definition,
+#'   peaks in `x` can match up to two peaks in `y` hence peaks in the returned
+#'   matrices might be reported multiple times. Note that if one of
+#'   `xPrecursorMz` or `yPrecursorMz` are `NA` or if both are the same, the
+#'   results are the same as with [joinPeaks()]. To calculate GNPS similarity
+#'   scores, [gnps()] should be called on the aligned peak matrices (i.e.
+#'   `compareSpectra` should be called with `MAPFUN = joinPeaksGnps` and
+#'   `FUN = MsCoreUtils::gnps`).
+#'
 #' @section Implementation notes:
 #'
 #' A mapping function must take two numeric matrices `x` and `y` as input and
@@ -238,16 +252,24 @@ NULL
 #' @param tolerance `numeric(1)` defining a constant maximal accepted difference
 #'     between m/z values of peaks from the two spectra to be matched/mapped.
 #'
-#' @param type For `joinPeaks`: `character(1)` specifying the type of join that
-#'     should be performed. See function description for details.
+#' @param type For `joinPeaks` and `joinPeaksGnps`: `character(1)` specifying
+#'     the type of join that should be performed. See function description for
+#'     details.
 #'
 #' @param x `matrix` with two columns `"mz"` and `"intensity"` containing the
 #'     m/z and intensity values of the mass peaks of a spectrum.
 #'
+#' @param xPrecursorMz for `joinPeaksGnps`: `numeric(1)` with the precursor m/z
+#'     of the spectrum `x`.
+#'
 #' @param y `matrix` with two columns `"mz"` and `"intensity"` containing the
 #'     m/z and intensity values of the mass peaks of a spectrum.
 #'
-#' @param ... option parameters.
+#' @param yPrecursorMz for `joinPeaksGnps`: `numeric(1)` with the precursor m/z
+#'     of the spectrum `y`.
+#'
+#' @param ... optional parameters passed to the [MsCoreUtils::join()] function.
+#'
 #' @return
 #'
 #' All functions return a `list` of elements `"x"` and `"y"` each being a two
@@ -257,7 +279,9 @@ NULL
 #' `y` input matrix have m/z and intensity values of `NA` in the result matrix
 #' for `y` (and *vice versa*).
 #'
-#' @author Johannes Rainer
+#' @author Johannes Rainer, Michael Witting
+#'
+#' @seealso [gnps()]
 #'
 #' @importFrom MsCoreUtils join ppm
 #'
@@ -286,9 +310,43 @@ NULL
 #' ## Right join: keep all peaks from y and those from x that match. Using
 #' ## a constant tolerance of 0.01
 #' joinPeaks(x, y, tolerance = 0.01, type = "right")
+#'
+#' ## GNPS-like peak matching
+#'
+#' ## Define spectra
+#' x <- cbind(mz = c(10, 36, 63, 91, 93), intensity = c(14, 15, 999, 650, 1))
+#' y <- cbind(mz = c(10, 12, 50, 63, 105), intensity = c(35, 5, 16, 999, 450))
+#' ## The precursor m/z
+#' pmz_x <- 91
+#' pmz_y <- 105
+#'
+#' ## Plain joinPeaks identifies only 2 matching peaks: 1 and 5
+#' joinPeaks(x, y)
+#'
+#' ## joinPeaksGnps finds 4 matches
+#' joinPeaksGnps(x, y, pmz_x, pmz_y)
+#'
+#' ## with one of the two precursor m/z being NA, the result are the same as
+#' ## with joinPeaks (with type = "left").
+#' joinPeaksGnps(x, y, pmz_x, yPrecursorMz = NA)
 joinPeaks <- function(x, y, type = "outer", tolerance = 0, ppm = 10, ...) {
-    map <- join(x[, 1], y[, 1], type = type, tolerance = tolerance, ppm = ppm)
+    map <- join(x[, 1], y[, 1], type = type, tolerance = tolerance,
+                ppm = ppm, ...)
     list(x = x[map$x, , drop = FALSE], y = y[map$y, , drop = FALSE])
+}
+
+#' @export
+#'
+#' @importFrom MsCoreUtils join_gnps
+#'
+#' @rdname joinPeaks
+joinPeaksGnps <- function(x, y, xPrecursorMz = NA_real_,
+                          yPrecursorMz = NA_real_, tolerance = 0,
+                          ppm = 0, type = "outer", ...) {
+    map <- join_gnps(x[, 1L], y[, 1L], xPrecursorMz = xPrecursorMz,
+                     yPrecursorMz = yPrecursorMz, tolerance = tolerance,
+                     ppm = ppm, type = type, ...)
+    list(x = x[map[[1L]], , drop = FALSE], y = y[map[[2L]], , drop = FALSE])
 }
 
 #' @importFrom MsCoreUtils localMaxima noise refineCentroids